little cleanup

+++ src/audio/AudioConfig.h
@@ -27,7 +27,7 @@
 class AudioConfig {
 public:
     uint_least32_t frequency; // sample rate (Hz)
-    uint_least32_t depth;     // bit depth
+    uint8_t        depth;     // bit depth
     uint8_t        channels;  // audio channels
     uint16_t       bufSize;   // sample buffer size
